Dynamic loader adjustments

+++ src/converters/BaseConverter.java
@@ -3,7 +3,7 @@ package converters;
 public interface BaseConverter {
  
     /* Da medida específica para o padrão do tipo (metro -> tipo) */
-    float fromBasicUnit(float value);
+    double fromBasicUnit(double value);
     /* Do padrão do tipo para a medida específica (tipo -> metro) */
-    float toBasicUnit(float value);
+    double toBasicUnit(double value);
 }
